userlist user update
